refactor: Move compiler and properties feature libs to one workspace lib
