+++ packages/compiler/drivers/astro/component.ts
