+++ packages/compiler/package.json
@@ -1,0 +1,15 @@
+{
+  "name": "@kitae/compiler",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "prettier": "^2.8.7"
+  }
+}
