+++ packages/compiler/properties/default-properties.ts
