+++ packages/compiler/properties/index.ts
