+++ packages/compiler/properties/renderer/ComponentColorsProperty.ts
