+++ packages/compiler/properties/renderer/ComponentRoundedProperty.ts
