+++ packages/compiler/properties/renderer/ComponentSpacingProperty.ts
